+++ embedded-graphics/src/prelude.rs
@@ -8,6 +8,7 @@ pub use crate::{
     pixel_iterator::{IntoPixels, PixelIteratorExt},
     pixelcolor::{raw::RawData, GrayColor, IntoStorage, PixelColor, RgbColor},
     primitives::{ContainsPoint, Primitive},
+    style::StyledPrimitiveAreas,
     transform::Transform,
     DrawTarget, Drawable, Pixel,
 };